feat: support decimal question scores

+++ src/lib/domains/exam/shared/contracts/start-exam.ts
@@ -6,7 +6,7 @@ export const examQuestionSchema = z.object({
 	type: z.enum(['multiple', 'short', 'essay']),
 	content: z.string(),
 	options: z.array(z.string()),
-	points: z.number().int().positive(),
+	points: z.number().positive(),
 	answer: z.string()
 });
